fixed clippy errors
Former-commit-id: 0cd00dc1ba5cf7b03d945405f5116dfaa3a098b4

+++ flow-core/examples/read.rs
@@ -49,9 +49,9 @@ fn main() {
         if (counter % 10000) == 0 {
             let elapsed = start.elapsed().as_millis() as f64;
             if elapsed > 0.0 {
-                println!("{} reads/sec", (counter as f64) / elapsed * 1000.0);
-                println!("{} reads/frame", (counter as f64) / elapsed * 1000.0 / 60.0);
-                println!("{} ms/read", elapsed / (counter as f64));
+                println!("{} reads/sec", (u64::from(counter)) / elapsed * 1000.0);
+                println!("{} reads/frame", (u64::from(counter)) / elapsed * 1000.0 / 60.0);
+                println!("{} ms/read", elapsed / (u64::from(counter)));
             }
         }
     }
